move to more intuitive directories; change name of `imshow_quantities`
