+++ chromatic/rainbows/visualizations/__init__.py
@@ -1,0 +1,5 @@
+from .colors import *
+from .imshow import *
+from .plot import *
+from .animate import *
+from .diagnostics import *
