+++ chromatic/tests/test_rainbow_normalize.py
@@ -24,4 +24,4 @@ def test_normalize(plot=False):
             r.fluxlike["relative-uncertainty"] = r.uncertainty / r.flux
             assert np.all(np.isclose(r.uncertainty / r.flux, 1 / snr, rtol=0.1))
             if plot:
-                r.imshow_fluxlike_quantities(maxcol=4)
+                r.imshow_quantities(maxcol=4)
